+++ index.js
@@ -1,23 +1,23 @@
 'use strict';
-const path = require('path');
-const fs = require('fs');
-const LgWebOsDevice = require('./src/lgwebosdevice.js');
-const ImpulseGenerator = require('./src/impulsegenerator.js');
-const CONSTANTS = require('./src/constants.json');
+import { join } from 'path';
+import { mkdirSync, existsSync, writeFileSync } from 'fs';
+import LgWebOsDevice from './src/lgwebosdevice.js';
+import ImpulseGenerator from './src/impulsegenerator.js';
+import { PluginName, PlatformName } from './src/constants.js';
 
 class LgWebOsPlatform {
 	constructor(log, config, api) {
 		// only load if configured
 		if (!config || !Array.isArray(config.devices)) {
-			log.warn(`No configuration found for ${CONSTANTS.PluginName}`);
+			log.warn(`No configuration found for ${PluginName}`);
 			return;
 		};
 		this.accessories = [];
 
 		//check if prefs directory exist
-		const prefDir = path.join(api.user.storagePath(), 'lgwebosTv');
+		const prefDir = join(api.user.storagePath(), 'lgwebosTv');
 		try {
-			fs.mkdirSync(prefDir, { recursive: true });
+			mkdirSync(prefDir, { recursive: true });
 		} catch (error) {
 			log.error(`Prepare directory error: ${error.message ?? error}`);
 			return;
@@ -67,8 +67,8 @@ class LgWebOsPlatform {
 					];
 
 					files.forEach((file) => {
-						if (!fs.existsSync(file)) {
-							fs.writeFileSync(file, '');
+						if (!existsSync(file)) {
+							writeFileSync(file, '');
 						}
 					});
 				} catch (error) {
@@ -80,7 +80,7 @@ class LgWebOsPlatform {
 				try {
 					const lgWebOsDevice = new LgWebOsDevice(api, device, keyFile, devInfoFile, inputsFile, channelsFile, inputsNamesFile, inputsTargetVisibilityFile);
 					lgWebOsDevice.on('publishAccessory', (accessory) => {
-						api.publishExternalAccessories(CONSTANTS.PluginName, [accessory]);
+						api.publishExternalAccessories(PluginName, [accessory]);
 						log.success(`Device: ${host} ${deviceName}, Published as external accessory.`);
 					})
 						.on('devInfo', (devInfo) => {
@@ -130,6 +130,6 @@ class LgWebOsPlatform {
 	}
 }
 
-module.exports = (api) => {
-	api.registerPlatform(CONSTANTS.PluginName, CONSTANTS.PlatformName, LgWebOsPlatform, true);
+export default (api) => {
+	api.registerPlatform(PluginName, PlatformName, LgWebOsPlatform, true);
 };